+++ lib/render.go
@@ -7,13 +7,19 @@ import (
 	"bytes"
 	"fmt"
 	"strings"
+	"regexp"
+	"os"
+	"html"
 
-	chroma "github.com/alecthomas/chroma/formatters/html"
+	chroma "github.com/alecthomas/chroma"
+	chromahtml "github.com/alecthomas/chroma/formatters/html"
+	chromalexers "github.com/alecthomas/chroma/lexers"
+	chromastyles "github.com/alecthomas/chroma/styles"
 	"github.com/yuin/goldmark"
 	highlighting "github.com/yuin/goldmark-highlighting"
 	"github.com/yuin/goldmark/extension"
 	"github.com/yuin/goldmark/parser"
-	"github.com/yuin/goldmark/renderer/html"
+	markhtml "github.com/yuin/goldmark/renderer/html"
 )
 
 func Markdown(source string, args []string) string {
@@ -38,8 +44,8 @@ func Markdown(source string, args []string) string {
 				highlighting.WithStyle(style),
 				highlighting.WithGuessLanguage(true), // this makes sure lines without language dont look bad! re:(^```$)
 				highlighting.WithFormatOptions(
-					chroma.WithLineNumbers(false),
-					chroma.WithPreWrapper(&preWrapStruct{}),
+					chromahtml.WithLineNumbers(false),
+					chromahtml.WithPreWrapper(&preWrapStruct{}),
 				),
 			))
 		}
@@ -51,8 +57,8 @@ func Markdown(source string, args []string) string {
 			parser.WithAutoHeadingID(),
 		),
 		goldmark.WithRendererOptions(
-			html.WithXHTML(),
-			html.WithUnsafe(),
+			markhtml.WithXHTML(),
+			markhtml.WithUnsafe(),
 		),
 	)
 
@@ -76,7 +82,7 @@ func (p *preWrapStruct) Start(code bool, styleAttr string) string {
 	styleAttr = strings.TrimSpace(styleAttr) // this param has spaces sometimes
 
 	if strings.HasPrefix(styleAttr, `style="`) {
-		newStyle := styleAttr[:len(styleAttr)-1] + `;overflow-x: auto"`
+		newStyle := styleAttr[:len(styleAttr)-1] + `; overflow-x: auto"`
 		fmt.Fprintf(w, start, newStyle)
 	} else {
 		// styleAttr doesn't start with 'style='
@@ -88,4 +94,52 @@ func (p *preWrapStruct) Start(code bool, styleAttr string) string {
 
 func (p *preWrapStruct) End(code bool) string {
 	return `</code></pre>`
+}
+
+
+var Code = regexp.MustCompile(`(?s)<pre><code[^>]*>.+?</code></pre>`)
+var LangRe = regexp.MustCompile(`<code class="language-([^"]+)">`)
+
+func Chroma(htmlsource string, style string) string {
+	f := chromahtml.New(
+		chromahtml.WithLineNumbers(false),
+		chromahtml.WithPreWrapper(&preWrapStruct{}),
+	)
+
+	return Code.ReplaceAllStringFunc(htmlsource, func (source string) string {
+		pre_start := strings.IndexByte(source, '>') + 1
+		code_start := strings.IndexByte(source[pre_start:], '>') + pre_start + 1
+		code_end := len(source) - 13 // 13 is len('</code></pre>')
+		code := html.UnescapeString(source[code_start:code_end])
+
+		// get lexer
+		m := LangRe.FindStringSubmatch(source)
+		var lexer chroma.Lexer
+		if len(m) > 1 {
+			lexer = chromalexers.Get(m[1])
+		}
+		if lexer == nil {
+			lexer = chromalexers.Analyse(code)
+		}
+		if lexer == nil {
+			lexer = chromalexers.Fallback
+		}
+
+		// get style
+		s := chromastyles.Get(style)
+		if s == nil {
+			s = chromastyles.Fallback
+		}
+
+		// tokenize
+		it, err := lexer.Tokenise(nil, code)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "Exception: %v\n", err)
+			return source
+		}
+
+		var b bytes.Buffer
+		f.Format(&b, s, it)
+		return b.String()
+	})
 }